[Feature] Optimize floors visited for least travel time
This introduces the functionality for the elevator to choose the order
in which to visit floors to ultimately reduce the total travel time as
well as the time to wait in the elevator. For example, given a set of
floors 1, 3, 12, 15, 19 (where 12 is the starting point) the elevator
will choose to travel 12, 15, 19, 3, 1. In the example given, the
shortest total path is to travel upwards first then downwards. Also note
that the order of the floors is such that the elevator visits each floor
in either direction before visiting the highest and lowest floors.